use Store object instead
