+++ CHANGELOG-LATEST.md
@@ -2,3 +2,4 @@
 
 - `purchaserInfo` renamed to `customerInfo`
 - Renamed `StoreProduct`, `StoreTransaction` and `StoreProductDiscount`. `package.product` to `package.storeProduct`
+- Store enum has been moved into store.dart
