+++ lib/models/entitlement_info_wrapper.dart
@@ -1,4 +1,5 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'store.dart';
 
 part 'entitlement_info_wrapper.freezed.dart';
 part 'entitlement_info_wrapper.g.dart';
@@ -19,32 +20,6 @@ enum PeriodType {
 
   /// If the period type couldn't be determined.
   unknown
-}
-
-/// Enum of supported stores
-enum Store {
-  /// For entitlements granted via Apple App Store.
-  @JsonValue('APP_STORE')
-  appStore,
-
-  /// For entitlements granted via Apple Mac App Store.
-  @JsonValue('MAC_APP_STORE')
-  macAppStore,
-
-  /// For entitlements granted via Google Play Store.
-  @JsonValue('PLAY_STORE')
-  playStore,
-
-  /// For entitlements granted via Stripe.
-  @JsonValue('STRIPE')
-  stripe,
-
-  /// For entitlements granted via a promo in RevenueCat.
-  @JsonValue('PROMOTIONAL')
-  promotional,
-
-  /// For entitlements granted via an unknown store.
-  unknownStore
 }
 
 /// Enum of ownership types
