+++ lib/object_wrappers.dart
@@ -8,6 +8,7 @@ export 'models/offerings_wrapper.dart';
 export 'models/package_wrapper.dart';
 export 'models/payment_discount.dart';
 export 'models/purchases_configuration.dart';
+export 'models/store.dart';
 export 'models/store_product_discount.dart';
 export 'models/store_product_wrapper.dart';
 export 'models/store_transaction.dart';
